change methods to private
